Rewrote a good part of the font code:
- surface->DrawString() now takes UTF-8 strings.
- Fonts do not need to be loaded completely any longer. The core keeps
an array of surfaces that grows when new glyphs are loaded. Information
about the glyphs is kept in a balanced binary tree using the Unicode
index of the glyph as the key.
- Kerning tables are not any longer in the core font struct. We now use
a callback from the core to the font provider and it can choose not to
implement it all.

Things left to do:
- Add some checks so we don't crash on very large fonts.
- Let the font provider specify the surface format and blitting flags.
- Hook into font destruction and release the FT2 face.
- Add DrawGlyph method to surface so we can use it from the GDK backend
or from a Pango module.