Added IDirectFB->WaitIdle() that waits for the graphics card being idle.
df_dok calls WaitIdle before and after each benchmark.
Now we see the real performance of the Savage3D.
Besides that e.g. the blitting benchmark takes 9 seconds
instead of 3 seconds defined in df_dok.c because it takes
the hardware 6 additional seconds to complete all the buffered
instructions. In this time (6 seconds!) the application can
use the CPU for other things ;-)