Do not use "-g" as default, add "-g3" to CXXFLAGS if in debug mode.
Added new function GfxCard->FlushTextureCache: after the video memory
has been written to by the CPU (e.g. modification of a texture) make
sure the accelerator won't use cached texture data. This new function
is called each surface_unlock (we should check if it was a soft lock
to video memory).

Implemented FlushTextureCache in Matrox driver, because font problems
are solved if any TEXORGn register is written to. Maybe this is just
a coincidence, but it works ;-)