Added option "force-windowed" which makes SetCooperativeLevel return DFB_ACCESSDENIED for other levels than DFSCL_NORMAL.
df_andi, df_particle and pss don't abort anymore
if setting the cooperative level fails.

Check for 'argc <= 1' instead of 'argc == 1' in df_dok,
because 'main()' can be called with argc set to 0 ;-)

Moved global variables of df_window into the main function,
removed layer enumeration that did not make any sense because
the IDirectFBDisplayLayer has only been constructed for the primary
layer, and that is DLID_PRIMARY just passed to GetDisplayLayer.