Added macro 'CAUTION' to "coredefs.h" that should be used when unimplemented features or unsolved problems have a serious impact on further execution, e.g. crashes or wrong behaviour.
Added surface resize handling code in 'IDirectFBSurface_listener'.
If the surface is a sub surface it's area will become the intersection
of the old area and the new surface extents. Prints caution message if
there is no intersection or the clipping rectangle is out of bounds.

Disabled code in 'IDirectFBWindow->Resize()' that is unneeded now, because
surface resizing is now handled in 'IDirectFBSurface_listener'.

Removed 'memcmp()' in 'IDirectFBSurface->SetClip()', it's always marked as
modified now, I hope some benchmarking will show which way is the best,
but this mostly depends on how often 'SetClip()' is called with the same clip.

Fixed typo or braino in listener attached by 'state_set_destination()'.

Changes above not fully tested, will test GDK backend soon.