Internal headers (all except "directfb.h") are now installed to "$prefix/include/directfb-internal". Files from "include", that are "directfb_internals.h" and "directfb-version.h", will be installed directly into this directory along with "idirectfb.h" from "src". Other headers from "src" will be in the same subdirectory as in the "src" tree. This way the '#include's are the same whether a module is built within or outside of the DirectFB source tree. First step to make it easier to write third party DirectFB modules in a seperate package.
There is a new pkgconfig file "directfb-internal.pc" which will
tell the third party configure script where the internal headers
are installed to. It will output "$prefix/include/directfb-internal"
when pkgconfig is called with "--cflags".

Next thing to do is version checking on modules, because it probably
will crash when using a module compiled with 0.9.4 is used with 0.9.5.