Added 'IDirectFBImageProvider->GetImageDescription' which fills out a 'DFBImageDescription', it contains a 'DFBImageCapabilities' field with the flags 'DICAPS_ALPHACHANNEL' and 'DICAPS_COLORKEY' for now. Additionally the Image Description contains the colorkey if DICAPS_COLORKEY is set, e.g. when loading a transparent GIF.
The GIF loader no longer generates an alphachannel from the colorkey,
because it turned the transparent color into black.

If you want to have transparent GIFs,
use colorkeying which is way faster anyway.


TODO: Turn 'IDirectFBSurface->SetSrcColorKey()' into a pixelformat
independent function that is called on the source for keyed blitting
to a destination, currently the colorkey of the source is set at the
destination in the source's pixelformat.