Important API change:
IDirectFBSurface::SetSrcColorKey has to be called on the source now,
when blitting to a surface which has source color keying enabled.

For IDirectFBSurface::SetSrcColorKey and IDirectFBSurface::SetDstColorKey:
The color key is now specified by three values each 8bit: r, g and b.
They will be converted to the pixelformat of the surface automatically.