Adding sophisticated keycode/keysymbol handling to DirectFB for more sophisticated applications like XDirectFB and the new keymap stuff in GDK. Writing input drivers for keyboards will also be easier.
There are to types of key mappings, a basic one (the one before) which is
suited for games and an advanced one for applications which need full
coverage of the Unicode 3.x character space and the ability to query the
keyboard layout.

The DFBInputDeviceKeyIdentifier is now contained in "event.key_id" instead
of "event.keycode". "event.key_code" is the hardware keycode which is usually
not of interest.

A new enum DFBInputDeviceKeySymbol is a super-set of the Unicode 3.x character
space which is compatible to Latin-1, so "key_ascii" has been removed in favor
of "event.key_symbol".

DirectFB header files (which include "directfb_keyboard.h" now) are installed
to "$prefix/include/directfb/". Applications honoring the CFLAGS within the
pkg-config file should have no problem with that.

Updated all copyright headers.

The keymap stuff is still work in progress, but the functionality to have
all applications running again already exists.


Short porting instructions:
- Replace keycode by key_id and DIKC by DIKI
  and/or
  replace keycode by key_symbol and DIKC by DIKS
- Note that the remote control keys have been removed from the
  basic mapping to reduce overhead
- Note that the cursor keys in the advanced mapping have a DIKS_CURSOR prefix
- Note that the letters have DIKS_SMALL and DIKS_CAPITAL prefixes