Added keymap support to input core.
Added a driver function to retrieve a complete keymap entry for a keycode.

Keyboard driver event thread is much simpler now, it passes the hardware
keycode and if it has been pressed or released to the input core via
dfb_input_dispatch() and the input core looks up symbol and identifier and
handles modifier/lock masks. The keyboard driver then simply writes back
the lock state to the LEDs.

Added DFBInputDeviceKeymapEntry that will soon be useful with some API calls.