Slightly improved description of IDirectFBFont API.
Create glyph surfaces with with a height of ascender - descender
instead of using font->height.

Don't bail out if setting the encoding failed in the FreeType2 font
provider. Users might want to use symbol fonts and should be allowed
to do so (they just need to know what they're doing).