Added experimental kernel device for a much better implementation of the fusion concepts. For now FusionRef is implemented in kernel space. The old implementation using semaphores doesn't work properly because undo counters are hold per process id. But the reference should be valid for all threads (processes with the same fusion id).
You have to apply the patch to test it, configure checks for linux/fusion.h
and uses the old implementation if the header isn't found.

The patch is here for quick testing. Will add a new cvs module with all
new files directly checked in and a patch for the modified.

The new implementation should be even faster as there is only one system
call per fusion_ref_* user space library call.

The undo stuff is working as supposed now (per fusion id),
but fusion_ref_zero_lock() is not implemented in kernel space yet.