Added RPC mechanism to let fbdev ioctls be executed by the master only. DirectFB now only detaches from its tty if it's a real VT and only if vt switching is enabled. This looks like a clean and working solution.
The only limitation left is that the master must not call DirectFB
functions that do fbdev ioctls within a thread that was created by the
application, but only if the application created any thread before DirectFB
got initialized. And of course this only applies to master processes started
from a real VT.

Added fusion_is_shared(void*) which returns true if the pointer specified
points to shared memory.