Fixed dead lock in write to stderr ;)
Scenario: OpenGL app locked the graphics card to do OpenGL stuff and
writes debugging messages to stdout or stderr. The DFBTerm it's running
in reads these debug messages and starts to render them waiting for the
garphics card lock. But if the io buffer in the kernel is full the write
of the OpenGL app blocks while it doesn't allow the DFBTerm to acquire the
lock.

Solution: Made graphics card lock a Fusion Property, like the lock for
display layers (shared/exclusive access). DirectFB graphics operations
are handled by the software driver now if an OpenGL context is locked.
That's not the best solution yet, but at least the dead lock is dead.

Fortunately moving a window over an OpenGL window always uses the hardware
because the lock is synchronized via the blocking surface buffer locks.
If a window is moved not touching the OpenGL window it gets slow ;(