Enhanced functionality of IFusionSoundPlayback::Continue().
...

The playback will begin at the position where it stopped, either
explicitly by Stop() or by reaching the stop position.

If the playback has never been started it uses the default start and
stop position which means non-looping playback from the beginning
to the end.

It returns without an error if the playback is running. This can be used
to trigger playback without having to check if it's already running,
similar to simple playback via IFusionSoundBuffer::Play(), but rejects
multiple concurrent playbacks.