Added IDirectFBEventBuffer::WakeUp().
Wakes up any thread waiting for events in this buffer.
This method causes any WaitForEvent() or WaitForEventWithTimeout() call
to return with DFB_INTERRUPTED.
This method should be used rather than sending wake up messages which
may pollute the queue and consume lots of CPU and memory compared to
this 'single code line method'.