Create a new process group early during initialization.
Send SIGTRAP to process group leader instead of the calling process
in DFB_ASSERT and DFB_BREAK. Otherwise the signal would be blocked
if the assertion or break happens in a thread, e.g. an input thread.

Call pause() after sending the signal, otherwise execution continues
until the signal is caught.