Hot fixed rare segfaults which occur if a valid shared memory pointer is derefenced but the address lies in a new region which hasn't been mapped locally yet.
Added "bool fusion_shmalloc_cure( const void *ptr )" which is called by the
signal handler to recover such segfaults by remapping the file immediately.
It returns true if the segfault is cured and the signal handler may drop it.

This is a hack until there's a better solution, maybe via Fusion Kernel Device.