Further finalization of the new layer code.
The primary region of a layer is no longer created and activated
during creation of the context.

IDirectFB::GetDisplayLayer() no longer causes the layer to be shown or
its surface to be created. Even IDirectFBDisplayLayer::SetCoooperativeLevel()
and IDirectFBDisplayLayer::SetConfiguration() no longer cause that.

A layer is allocated and shown upon:
1) IDirectFBDisplayLayer::CreateWindow()
2) IDirectFBDisplayLayer::GetSurface()
3) IDirectFB::CreateSurface() <- if primary & fullscreen

Fixed failing assumption for exclusive access to additional layers.
Fixed missing context and region destruction after releasing the layer.