Store the serial of the most recent accelerated graphics operation that has been queued for writing to a surface buffer.
dfb_layer_region_flip_update() is no longer waiting for an idle accelerator,
if the buffer that is becoming the front has or had a write operation queued.

Instead it waits exactly for the last operation that is or was about to write
to this buffer ;)

If WaitSerial() is not supported by the driver and/or hardware, the fallback,
of course, is to wait for each and every operation to finish.