Removed StReT stuff from default window manager module.
Forked a new window manager module with the new StReT stuff.

The new window manager consists of the wm module that is plugged into
DirectFB and a library that contains the wm implementation. This library
can be linked into programs, like in the "test_color" example which
changes the global window border color. This test will be removed and
replaced by a more sophisticated configuration system.