Added DLCAPS_SOURCES which indicates that the layer supports multiple sources that can be selected, e.g. from frame buffer (surface) or from hardware decoders (using buffers outside of the frame buffer).
Added DFBDisplayLayerSourceID with pre-defined DLSID_SURFACE.

Added number of supported sources to DFBDisplayLayerDescription.

Added DFBDisplayLayerSourceDescription containing the ID and a name.

Added DLCONF_SOURCE along with a field in DFBDisplayLayerConfig to
select the source to be used.

Added IDirectFBDisplayLayer::GetSourceDescriptions().

Added InitSource() to the layer driver API.

Added source_id to CoreLayerRegionConfig
and CLRCF_SOURCE_ID to CoreLayerRegionConfigFlags.

Added layer sources to dfbinfo.