Added DLCAPS_ALPHA_RAMP:
Alpha values for formats with one or two alpha bits
can be chosen, i.e. using ARGB1555 or ARGB2554 the
user can define the meaning of the two or four
possibilities. In short, this feature provides a
lookup table for the alpha of these formats.


Added "__u8 alpha_ramp[4]" to DFBDisplayLayerConfig along
with DLCONF_ALPHA_RAMP in DFBDisplayLayerConfigFlags:

Alpha values for 1 or 2 bit lookup.
See description of DLCAPS_ALPHA_RAMP.
Either all four values or the first and the
last one are used, depending on the format.
Default values are: 0x00, 0x55, 0xaa, 0xff.