Applied path from Sylvain Meyer <sylvain.meyer@worldonline.fr>, thank you!
        Hi,

    I did a small patch to fix minor issues with i830 gfxdriver. Now the
driver can more nicely live with the linux intelfb driver. Now it can be
used even when intelfb has accel set to 1, just in configuring in
/etc/fb.modes "accel false" to disable console acceleration when
DirectFB is working. To work, the last intelfb patch is needed. I just
sended it to linux-fbdev mailing list and Andrew Morton and it should be
in the next mm series of linux kernels.
    I also fixed the ovlSetColorAdjustment function and now brightness,
contrast and saturation settings are enabled.

Regards
Sylvain