Overhaul of LEVELS, OPACITY and primary ALPHACHANNEL features which were previously only partially implemented and did not work together.
The video overlay and the primary layer now both support the OPACITY option
and either can be on top.  The order is determined by the 'level' assigned
to the video layer.  The primary also supports an alpha channel which is
useful when the video layer is positioned behind.

The old DFB_CLE266_UNDERLAY environment variable switch has been removed.