DIKI_ALTGR is gone. The key right to the space bar is the right alt key. No matter which map is loaded, the identifiers are just named hardware keys, where you really specify the physical entity on your keyboard. In this case it's always DIKI_ALT_R, no matter if it's mapped to DIKS_ALTGR or DIKS_ALT. On standard keyboards it's also always the same hardware key code, no matter if you by a German keyboard with AltGr or a U.$. one with Alt.
Changed the input core logic for keeping track of modifiers. If DIKI_ALT_R
is pressed, also look at the symbol to decide if DIMM_ALT or DIMM_ALTGR is
to be added/removed.

Now application developers don't need to check for both DIKI_ALTGR and
DIKI_ALT_R if they want to check for the key right next to the space bar :)