Moved clipping from dfb_clip_stretchblit() to gStretchBlit() doing all clipping, phase and offset calculations in place, well, handling the phase of clipped stretching at all (initial .16 value for iteration) hasn't been there before.
Before that, clipping suffered from rounding and off-by-one errors due to
going back to integer in between.

Accesses out of surface bounds and jumping images are gone. Jumping was
visible when the same image was stretched with the same source and
destination rectangle each time, but with randomly changing clip rectangles.