DGIFF - DirectFB Glyph Image File Format
Added a new font module and a tool called mkdgiff to generate .dgiff from .ttf or other.

Purpose of DGIFF is to offload as much as possible from runtime, especially during
application startup when loading the default or system fonts. mkdgiff uses FreeType2
to load the glyphs of the font and stores them in one or more ready to use font cache
rows, up to 2047 pixels wide right now.

The loader creates preallocated surfaces from the mmapped font file. The only copy ever
made would be when the surface is transfered to video memory for acceleration.

DGIFF supports different sizes of one face in a single file.

You can specify these options on the mkdgiff command line:

   -f, --format    <pixelformat>   Choose the pixel format (default A8)
   -s, --sizes     <s1>[,s2...]    Choose sizes to generate glyph images for