Initialize the heap info table to serve for the maximum pool size specified at creation of the pool. Previously, the initial table served for 4 MB heap. If there were more allocations, the heap info table was reallocated at the end of the heap, i.e. at 4 MB making it impossible to ever go back below the 4 MB heap size. Next barrier was 8 MB, 16 MB ...these are gone.
Unlink potentially remaining file from a previous session before creating
the shared memory file in tmpfs.

Removed memset() on the initial heap to take advantage of the actual allocation
of the pages happening when first accessed. If it's a new file, which it is
now, it will be zeroed anyways.

The maximum size specified during creation of a pool is now taken as an
additional size after the heap info table. Previously, it could happen that
the size was too small to host the fixed sized heap info table.