If trace support and shared libraries are enabled, for each library to install, also generate a file containing the output of "nm -n <that library>".
That's the command that is used at runtime via popen().

Now the code first checks if it can fopen() a file with the same name as the
one passed to "nm -n", but with an "nm-n." prefix.

This is both a performance improvement and a nice solution for (embedded) targets
without 'nm' being available or fast enough.