Use DFBUpdates to manage an update queue and do optimizations like taking the bounding box instead of multiple smaller regions if the number of pixels in the bounding box is not much higher than the total of the smaller ones.
Even this takes the number of smaller regions into account, because of the
increasing overhead per region.

Next to performance, the visual experience is enhanced by coalesced updates.

Another advantage is that when multiple regions are updated, first it renders
all regions and then it flips all regions. Previously, each time the flip was
done immediately for each rendered region.