Fix fusion_sync() timeout during XDirectFB shutdown.
Apart from other small changes, fusion_sync() returns DFB_OK in case
the select() call was interrupted (EINTR). Continuously sent signals
could have cause these timeouts, but I don't know if and why they would
be there.

It's still a hot fix, as it could be that the call is interrupted with
data available, couldn't it?

A proper solution could be to support fsync(fusion_fd), but what about
the timeout?