Added DWDESC_OPTIONS and DWDESC_STACKING to give initial values in the DFBWindowDescription using the new fields 'options' and 'stacking'.
Added DWDESC_PARENT to DFBWindowDescriptionFlags and 'parent_id' to
DFBWindowDescription. This can be used to associate a window to another,
but it's up to the WM to implement it. In SaWMan this makes the window
cover exactly the area of the parent window including scaling if needed.
Associated windows do not get focus or events themselves, but their parent.

Added IDirectFBWindow::SetKeySelection() that selects a mode for filtering
keys while being focused. The selection defines whether all, none or a
specific set (list) of keys is selected.

Added DFBWindowKeySelection featuring DWKS_ALL, DWKS_NONE and DWKS_LIST.

Added IDirectFBWindow::GrabUnselectedKeys() and UngrabUnselectedKeys()
to have a key collector window receiving all keys that the focused
window did not select.

Added DFBWindowEventFlags featuring DWEF_RETURNED so far, which means that
the event has been returned, i.e. has not been consumed by the original
recipient. The DFBWindowEvent field 'flags' has also been added.

Simplified internal dfb_layer_context_create_window() and dfb_window_create()
and replaced dfb_window_*grab*() by dfb_window_change_grab().

Changed IDirectFB::CreateSurface() behaviour: For primary surfaces in
windowed mode, only create a window with two buffers if DSCAPS_DOUBLE
is set, but not DSCAPS_FLIPPING.

Added DWDESC_PARENT to DFBWindowDescriptionFlags and 'parent_id' to
DFBWindowDescription. This can be used to associate a window to another,
but it's up to the WM to implement it. In SaWMan this makes the window
cover exactly the area of the parent window including scaling if needed.