Implemented StartDrawing/StopDrawing() for states.
When a surface is being rendered to, its graphics
state is set to DRAWING. Upon Flip() the state is
flushed.

Each sub surface interface has a reference to its
parent and a list of its children.

The new mechanism needs to be enabled via "startstop"
option.

StartDrawing/StopDrawing() can be implemented by the
graphics driver. e.g. to keep track dirty surfaces
and defer flipping if it would reveal other's drawing.