Revert "Flush the read cache only when a CPU read access is requested."
This reverts commit fd596f8f4c906d78ebbf1d710448aaef01cfb208.

If a CPU write access is requested but nothing is actually written to the
framebuffer the hardware will not flush the cache. The next CPU read access
will fail to explicitly flush since the GPU write flag has already been
cleared. So simply flush the read cache even on CPU write access.