Fix SetMode / Panning Bug
When a graphics driver has its own primary layer hooks implementation, and
then defers to fbdev for primarySetRegion(), dfb_fbdev_set_mode() is called
unnecessarily.

The original implementation in DFB1.0.0 would only call dfb_fbdev_set_mode()
if a flag indicated a change, however now that dfb_fbdev_set_mode() is in
the default case of a switch statement, it is always called regardless. This
has the side effect of resetting the panning, when features implemented in
the graphics driver are called such as setting opacity.