Implemented defered region realization by adding CLRSF_FROZEN which is set initially on every region that is created.
As long as CLRSF_FROZEN is set, all SetRegion() calls to the
driver (and buffer allocations due to it) are simply discarded.

Only when dfb_layer_region_flip_update() is called and the region
is enabled and active, the flag is removed and the region is either
realized or the config is just set if it was already/still realized.

That means from a driver's point of view: AddRegion(), SetRegion()
and either FlipRegion() or UpdateRegion() are called in a row for
the first time, with a buffer that is already initialized with the
proper content.

However, the driver should still not show the region in SetRegion(),
but in FlipRegion() or UpdateRegion(). This is because between the
SetRegion() and the FlipRegion() call, the locked buffer changes to
the correct one (only for multi buffer modes).

NOTE: This needs more testing, especially single buffered exclusive
contexts where the application might never call dfb_layer_region_flip_update().