Added "[no-]show-empty", init default borders, center unscalable layers.
Using "show-empty" makes SaWMan show the layer even if no window is visible.

Default borders (red/grey) are configured with thickness of 4. To disable
borders, use "border-thickness = 0".

Fix bug with reactivation of SaWMan's layer context's, e.g. when an exclusive
context is returned from.

In standard mode, center layers that cannot be scaled. Previously, it just
tried to scale the layer to fullscreen.