API change in IComaComponent::InitNotifications() and InitListeners().
Added context pointer argument which is registered with the listener or
notification dispatch callback if its entry in the array contains a
NULL context pointer.

Made array arguments const.

These two changes allow definition of static const tables of notification
dispatch callbacks where the context is only known during runtime and is
the same for all of these.