Added TI Davinci driver.
Supports OSD and Video with positioning, color key, opacity and alpha channel.

The driver uses the devmem system and opens all frame buffer devices itself.
This is required for extended ioctls like setting the OSD position, but also to
support alpha channels which are stored in a different frame buffer device. In
case of ARGB for the OSD, the alpha channel (3 bit!) is dithered during conversion.

Implemented surface pools for OSD and Video layer. This enables direct usage
of the frame buffers in case of RGB16 for OSD and UYVY for Video.