SH7722: Implemented DSRO_ANTIALIAS for all drawing operations.
The hardware provides "anti-aliased" lines by drawing a line
with doubled pixel width (minor axis). This should be drawn with
a transparency factor. Finally the original line is drawn on top
of this.

For filled rectangles and triangles, only the double width line is
drawn and then the shape is filled.

Fixed off-by-one in rectangle outlines.