DAVINCI: Real IRQ based synchronization, optimization for opaque filled rectangles.
Added C64X_FLAG_INTERRUPT to get an interrupt when the command is finished.

Function zero is NOP now, e.g. to generate an interrupt at the end of the queue,
without setting the interrupt generation bit in already queued commands.

davinci_c64x_wait_low() no longer does a usleep() leading to big performance loss,
when it's called often with almost nothing on the queue. If synchronization needs
to be done, it queues a NOP with Interrupt flag and calls C64X_IOCTL_WAIT_LOW to
wait inside the kernel for the end of the queue, and continue without a huge latency.

Because blended rectangles are much faster, use the blending function on DSP also
for non-blended filling if possible, i.e. alpha = 0xff.

Updated tests. Cleanups.