surface core: Fix a potential problem with deallocating buffers.
dfb_surface_buffer_lock() walks the buffer allocations with
fusion_vector_foreach (alloc, i, buffer->allocs) and if it doesn't
find a suitable allocation it will allocate one. It then tests
alloc against NULL in the error paths to determine if the
allocation was performed. If there are only unsuitable allocations
alloc will not be NULL even though a new allocation was made.
So use other means to keep track if allocation was made.

In the name of keeping the code more readable give
dfb_surface_buffer_write() the same treatment even though it's
not actually necessary there.